Add another binary stack interpolator
This plugin expects a binary stack. The non-black slices are used to
interpolate the black slices.

The principle: each non-black slice is subjected to the signed integer
distance transform (see below). For each black slice, then the transformed
next non-black slices on either side are weighted by z distance, added, and
the result is inverse-transformed.

The distance transform attaches the shortest distance to the boundary to
each pixel.

The _integer_ distance transform uses a modified Manhattan distance instead
of the Euclidean: the integer distance between two pixels is the smallest sum
of neighbour distances between the pixels, where horizontal and vertical
neighbours have distance 3, and diagonal neighbours have distance 4 (in 3d,
the 3d diagonals would be 5, but we only use 2d here). It is a relatively
good approximation to the Euclidean distance, with the benefit that the IDT
can be calculated very fast: the image has to be processed once for all
four directions North, East, South, West.

The signed IDT attaches a negative sign to the outside pixels (so that
all inside pixels have an sIDT value >= 0).