Introduce ij3d.IJ3dImageWindow
This is the proposed base class for our 3D stuff. As of now, it is
an ImageWindow, and as such shows up in ImageJ's window list. It
also supports the method getImagePlus(), which means that you can
Image>Duplicate... the window, creating an ImagePlus by doing so.

This feature does not work on all machines, though! On some boxes,
Bene and me experienced errors because of a lacking pbuffer support.

The plan now is to work around that, by requesting the focus to the
3D window, and getting a screenshot of the relevant part.