Start big rewrite of the modules
It is not necessary to rely on reflection. Instead, we rely on recursion.
At a later stage, this framework can be used to make the application
multi-threaded, and possibly clusterable.

To this end, vib.app.module.Module _is_ the end module, i.e. the entry
point into the protocol. A module calls all the modules it depends on, in
one of two modes: there are modules which operate on single images (like
Resample), and there are modules which operate on the complete set of
images (like AverageBrain).

Also, this introduces a separation between the options, which are set by
the user, and the state, which is set by the application.

This is a work-in-progress. At this moment, it does not work at all, since
only Module, AverageBrain and TransformImages are converted to the new
framework.