The Isosurface viewer now allows to record a movie.
Right-click -> Start recording starts movie recording. Right-click ->
Stop recording stops movie recording and displays the resultant stack,
which can be stored as .avi, for example.

This is done in the IsosurfaceUniverse via PickCallback: after each
rotation step the callback method is called, which grabs an image
of the ImageWindow3D and adds it to a stack.