First attempt to merge the Volume Renderer and Isosurface viewer.
There exists now a temporary package called vis3d, which contains the
universe (Image3DUniverse) and class called Content, which is the base
class for VoltexGroup and MeshGroup.
VoltexGroup is one Volume Renderer entity, MeshGroup is on Isosurface Viewer
entity.

vis3d will be moved to ij3d in the future.