Allow to choose a color for a volume.
The color field was previously hold by AxisRenderer (without being used) and
is now transferred to Axis2DRenderer. Here, the color is used to set colors
for the underlying geometry. The right texturing mode determines the resulting
colors, which is a combination of the geometry color and the texture color.
This is not perfect yet. The goal is to use the color specified by the user,
unless 'none' is selected. Then use the texturing color (which comes from the
indexed ImagePlus).