Allow to set the color of a volume.
As intended, the user can select a color, if he wants to apply it to the
volume, or select none, to use the colors of the ImagePlus.
Therefore, re-introduce a field in Volume.java which indicates which ColorModel
is to be used: grayscale, if a default b/w model should be applied, otherwise,
the colormodel of the ImagePlus is used.

In Axis2DRenderer, use the correct TexturingMode: REPLACE, if the no color is
selected, MODULATE otherwise.