Integration into ImageJ:
Duplicate, Crop and Copy of the 3D stuff now takes ROIs into account:

Therefore, made some structural changes in the ImageCanvas3D, ImageWindow3D
and Image3DUniverse:
Image3DUniverse creates an ImageCanvas3D at startup. The ImageWindow3D
is created out of the universe (in the show() method) and gets a reference
to it. Thus, it does not have to create an own Canvas3D, but rather uses
the ImageCanvas3D of the universe.

Finally: getImagePlus in ImageWindow3D sets the ROI of the ImageCanvas3D
before returning the result.