Changed the structure of the ij3d.Image3DUniverse:
Added the interface UniverseListener, which listens to transformations and
content add/remove.
The base class of the universe, DefaultUniverse, implements the fire...
methods, and the add/remove listener methods.
DefaultAnimatableUniverse extends the DefaultUniverse and adds methods for
animation.
Image3DUniverse finally extends DefaultAnimatableUniverse and implements the
add/remove Content methods.