Install_Java3D.java: new plugin which installs Java3D.
What it does: It reads java.home, which gives the directory of the
jre used by ImageJ. This is especially useful for windows, since the
windows version of ImageJ is optionally bundled to a own JRE.

The user has to download the java3d zip file and specify it in the
appearing dialog. This is then unzipped in the Java home dir, which should
place the files to appropriate folders.