Change the color changing process.
There exist now two color models: one is the index color model of
the original image (maybe encoding more than one channel), the other one
is a default grey color model of the selected images.

If the color model is not changed, and just the color has to be
updated, this is done by the TextureAttributes (in the case of voltex).

In the case of the isosurface, this is not necessary anyway, since the
color is set as color of geometry (which is fast enough).