Implemented recalculateGlobalMinMax() in Image3DUniverse.
This calculates the new global minimum and maximum point and centers the
universe according to that. It is called each time a content is added.

Unfortunately, this breaks the MouseBehaviors correct...() methods, which
previously multiplied global and local rotation into the local one, so
that further mouse behavior resulted in an expected transformation.

Therefore, this method is now commented out.