ImageWindow3D: toggle "HAND" tool with the Escape key
Like with Amira, the Escape key can be used to switch to the Hand tool
and back.

Unfortunately, the Escape key in ImageJ has already the meaning of
aborting a macro or plugin, but you are likely to want to change the
tool with the Escape key only if you are using the 3D Viewer
interactively anyway.

Also, it is easy to transfer this behaviour to another key.