Implemented saveTransform() in Image3DMenubar. Also fixed one bug which occurred with setTransform():
The local rotation group in Content also contains translational elements.
This is due to the mouse behavior: It first translates to the center, rotates
and translates back. This concatenation results in a matrix with translational
elements.

When setting a transform, previously I separated the matrix into translation
and rotation/scale. But I did it wrong, just taking the upper 3x3 matrix as
rotation: This matrix must always have a translational component:
Let R be the rotation matrix and c the center point. Then the translational
component of the rotation matrix is always equal to
| -r11*c1 - r12*c2 -r13*c3 + c1 |
| -r21*c1 - r22*c2 -r23*c3 + c2 |
| -r31*c1 - r32*c2 -r33*c3 + c3 |

The rest of the translational part of the specified matrix goes into the
local translation transform.