Psychomorph TEM Reader: major restructuring
The TEM Reader now contains a class TEM, which has conciser parsing
and ROI setting routines, and it learnt as well to add points, lines,
and write out the whole thing.

As a test, the class contains now a main() method, which takes two
existing TEMs (where the first is expected to contain more points
and lines), uses MLS to extrapolate the missing points of the second,
and writes out a new TEM file.