Handle failure to off-screen-render more gracefully
Java3D 1.5 has a RenderingErrorListener interface, and you can add
listeners to every Universe.  If you do not add one, the default error
handler will be called, which just exits the program.

Not what I want.  So I want to add a RenderingErrorListener.

Java3D 1.4 apparently does not have a RenderingErrorListener.  Therefore,
the code will neither compile nor run with Java3D 1.4.

Not what I want.

So here is an ugly workaround: we add a minimal implementation of
RenderingError, and the interface, so that it compiles with Java3D 1.4.

Then we get the Universe's method addRenderingErrorListener() at runtime (!)
and call it, if it exists.  And I thought we were past RTTI.  *sigh*

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>