Fix compilation of target 'all-jars'
This marks an important point: For the first time, I was unable to use
some sed foo to compile all classes in compatibility mode.

The culprit is vib/transforms/FastMatrixTransforms, which inherits
"FastMatrix inverse()" from FastMatrix, but should implement
"Transform inverse()" for the interface Transform it purports to
implement.

This is not clean object orientation, and therefore older Java refuses
to compile it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>