Begin of a general code cleaning and refactoring of the 3D things.
Note: This version may not compile or work.

This and the next commits are thought to refactor the 3D classes:
The goal is to have one Renderer class (instead of 3), and helper
classes to delegate the work, e.g.
GeometryCreator.java is going to create a geometry out of a specified
axis and an index within this axis.
AppearanceCreator.java is going to create the appropriate texture and
appearance, again out of a specified axis and index within that axis.

Geometry and Appearance are used in the renderer class to create a Shape3D,
which is then added to the scene graph.

The plans for Orthoslice are to let it inherit from the volume renderer class.

This particular commit now took much of the code of AxisRenderer and gave it
to the Axis2DRenderer, which is going to be the only renderer class.
AxisRenderer itself is going to become the GeometryCreator class.