Total restructure of content node.
The idea now is:
Content is no longer the abstract superclass of OrthoGroup, VoltexGroup and
MeshGroup. It fullfills now the task of creating and managing the local
transformations. A second task is to hold the attributes of ContentNode, which
is the new abstract superclass of OrthoGroup, etc.

This design has the benefit that a Content can now exchange its ContentNode,
thereby making it possible to change the display mode of a content, e.g.
switch from Volume to Surface etc.

Another side effect of the restructuring was to sort and re-use more of the
code.