Use calibration information in ComputeCurvatures and friends
A "FIXME" with this commit is that it currently assumes that
pixelWidth will have at least as small a value as pixelHeight
and pixelDepth.  I will fix in a later commit this so that the
default sigmas are based on the smallest of the three rather
than pixelWidth.

Other changes included in this commit:
  - Display progress properly in ComputeCurvatures_
  - More whitespace cleanup