Split the Image3DMenubar into two classes:
The Image3DMenubar now just contains the GUI components, i.e.
the menus and menu items. It also implements the ActionListener
interface, however, the executing functions are now implemented
in Executer.java.

The Executer also has a SliderAdjuster thread, which according to
Albert's TransparencyAdjuster is supposed to handle all the slider
updates.
The SliderAdjuster is abstract, so that each superclass can
implement the actual updating function (like setting the
transparency, etc).