Take the origin of the images into account when displaying images.
The origin is interpreted as the location where the lower left corner
of the volume should be placed. This might not coincide with the
intention behind the origin fields in the calibration, but it's good
enough for testing now.

The Geometry- and AppearanceCreator are adjusted so that they place
the volume to (xOrigin, yOrigin, zOrigin).

Adjust AppearanceCreator and GeometryCreator in a wa