Speed up undisplay(Cube) by giving a hint in which axis to find it.
When the cube's display method is called, the axis is given as a
parameter. The cube memorizes this axis and knows where it is, and
where it has to be removed from when undisplayed.

Also fixed code which takes care that the current cube is undisplayed
before it is displayed in a different direction.