When transforming, show always a rough representation of the volume.
ShapeContainer's axisSwitch has now 7 children: 6 of them hold the
shapes for a rough representation of the volumes (one for each of the
six directions), the 7th holds a detailed representation.

When a transformation is started, the detailed representation is
switched off and the rough representation is switched on. In this
way, it is also possibe to rotate a volume with many many shapes.

While the volume is transformed, the current axis is updated (for the
rough representation).

As soon as the transformation is finished, the rough representation is
switched off and the detailed representation of the corresponding axis
is loaded into the 7th child of the axisSwitch and switched on.