Avoiding the use of the CubeDataRecycler.
Holding the data in memory for recycling leads to
memory problems. By not using the CubeDataRecycler, I get
really good rendering for my 1024x1024x256 stack, for the
first time.

This version runs quite stable, there's no memory accumulation,
and using a window size of 512x512 gives good results.

The recycling mechanism may be good in general, but needs more
work: Restricting the number of stored elements, depending on
the amount of free memory, for example. So I will keep the
corresponding classes and work on it later.