A general change in the octree display mechanism:
VolumeOctree collects now (recursively) all the cubes which are
to be displayed in the next frame. During this process, all cubes which
will be displayed set a flag whether their data needs to be updated.
Also during this process, cubes which will not be displayed release
their data.

After collecting the cubes, all the ShapeGroups of the cubes are thrown
in one array, which is sorted according to the ShapeGroups' z-position.
Afterwards they are added to the scenegraph, one by one. Before
displaying a ShapeGroup, its Cube is asked if it is uptodate, and if
not, it is updated.

There are a few advantages of this behavior:
* It is very fast derived how many cubes have to be loaded. The sorting
  is done before the image loading. So only after rendering a few shapes
  the visible result should be very good (because we do ordered
  rendering). This is not yet implemented, since all the rendering is
  started from the Behavior class, so is done at once.

* It is known at the beginning, how many cubes will be needed. So one
  show a progress bar.

* The step which involves the loading of the images is no longer done by
  recursion through the cubes, but centralized in the VolumeOctree.
  Cancelling the update process should thus be easier. This is also not
  yet implemented.