Changed the structure of the scenegraph for octree volumes.
Now the axis switch has a BranchGroup before the OrderedGroup. When
the tree is updated, that BranchGroup is detached, and the shapes are
added to a new, non-live BranchGroup. This speeds up adding the shapes.
Additionally, one does not need to wrap the shapes in BranchGroups any
more, so ShapeGroup does no longer extend BranchGroup, but just holds a
reference to the corresponding shape.