Remove ShowAmiraNeuronTree
Nobody seems to use it, and it is the last user of Quick3dApplet, which
we want to get rid of in favor of Image3DUniverse.

If it is needed again, it should be easy to resurrect (and we will have
the test cases we lack now, to verify that the conversion was correct).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>