Lasso/Blow: support all image types, actually be useful
Now, Lasso and Blow will support all image types, with a special speedup
for byte and color processors.

Also, it will in fact adjust the selection _as long_ as you click the
mouse, instead of _after the first click_.  That means that when you
no longer press the mouse button, the selection is no longer adjusted,
and you can actually do useful things with the selection.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>