Name_Points: Better template handling (based on a patch by Gregory Jefferis)
There is no hard-coded default template file any more.

On startup, a template specified in ImageJ preferences
is used if possible.  If the user is holding down Alt
while starting the plugin they will be prompted to
select a new template, and that template will be set
as a default in the preferences.

The default template file can also be set and unset
with the "Set as Default" button, since on my setup
IJ.altKeyDown() is far too difficult to trigger :)

If there is a default template and the current image
appears to have no points file, the default set of
point names will be those from the template.

If there is no default template and the current image
appears to have no points file, then a single example
point name is added.