The Texture_By_Ref Plugin implements now PlugInFilter.
It loads the image provided to the PlugInFilter and allows directly
to draw on it.

The tricky part of this step was to create a BufferedImage from
a given byte[] array. I found finally code for that in the
ImageProcessor classes.

The plugin was moved to a newly created package textureByRef.