Make VoltexVolume editable.
Moved all the funcionality of EditableVolume to VoltexVolume;
VoltexVolume holds now byte[] arrays for the data, as well as
ImageComponent2D[] arrays for the texture images.

The loaders implement the set() and setNoCheck() methods so that
all perspecitves of the data are modified.

A ComponentCreator is responsible for the creation of ImageComponent2Ds
from the byte data. It is specialized as GreyComponentCreator and
ColorComponentCreator.

Volume_Editor uses the normal voltex/ classes, and no longer
textureByRef.EditableVolume and textureByRef.EditableVolumeRenderer.
These classes are no longer needed and will be removed in the following
commits.

Finally, Volume_Editor was adjusted to handle RGB stacks.