Changed the logic between PointList, PointListShape and PointListPanel;
This mainly concerns Content.java;

A PointList is now created first, and a reference to it is given to the
PointListShape and PointListPanel.

Both of these classes register at the PointList as PointListListeners,
and thus get informed about changes in the underlying PointList.
Therefore, the PointList itself can be used for operations on the list.