3D Viewer: Detect Java3D version, and if necessary, update.
ij3d.Install_J3D is a new class which can check the version of the
installed Java3D and install a new (hard-coded 1.5.2) version.

It downloads the zip-file from Java3D's homepage, according to the operating
system. These zip-files contain a file j3d-jre.zip, which contains the
important files. For windows and linux, j3d-jre.zip is just extracted
in the java.home directory (preserving the directory structure), on Mac,
j3d-jre.zip is extracted in a temporary folder, and the jars are then
copied directly to the first folder as given by java.ext.dirs.

The user is prompted for each file which existed before if it should be
overwritten.

On Mac, Jogl has to be installed in addition to Java3D. This is done
equivalently, and the jnilibs are copied directly to the same java
extension folder.

The actual check is done in ij3d.ImageJ3DViewer. However, in the case
where no Java3D was installed before, it used to crash. The reason was,
that ImageJ3DViewer had a static field to Image3DUniverse. When the JVM
loaded the ImageJ3DViewer, it tried to load the Image3DUniverse class at
the same time, without giving ImageJ3DViewer the chance to check for
Java3D.

To circumvent this behavior, the static reference was now deleted.