3D Viewer: Changed rotation behavior.
There is now a distinction between Universe.globalCenter, which is just
the midpoint of all displayed contents, and
ViewPlatformTransformer.rotCenter, which is
a) the center of the view, therefore changed by centerSelected(). It's 3
   values specify the transformation of centerTG.

b) at the same time the rotation center.

This guarantees that rotation is always around the center of the view.