3D Viewer: Changed the ordering of the ViewingPlatform's transform groups.
In particular, the TransformGroup is now the first one in the hierarchy,
which has the effect that zooming can always be done by moving in local
z-direction of the zooming TG.

Since translation should not change the center of rotation, this makes
rotations much easier: Having the center of rotation fixed to the center
of view, there is no back-translation required any more.

In ViewPlatformTransform, all the rotation methods which take a rotation
center as input parameters are still present, but the given point is
ignored. Rotation center is always the center of the view. The
corresponding methods will be removed in the next commit.