Minor synch issues solved, still some to go.
The ImageWindow3D$ImagePlusUpdater logic was wrong: the "update"
flag was being checked outside a synchronized block, thus making
it possible to slip through in some cases.

The OctreeBehavior.display method, which is private final,
should not be synchronized: it's launched from within a Runnable
executed by an ExecutorService, and the other synchronized method
is the one that submits the Runnable to the ExecutorService: makes
no sense at all to block calls to the launcher method. Why thread
it then?