Fixed recalculation of normals.
When initializing a GeometryInfo with a QuadArray, and then computing
normals, the quad array is first converted into indexed triangles. So
the normals are also indexed.
unindexify() does remove the indices, but converts the data back to quad
arrays, of course.

If using getGeometryArray() when only computeNormals() is called in the
GeometryInfo, the data is indeed converted back to quad arrays.
(see docs for GeometryInfo.getNormalIndices()).