Simple_Neurite_Tracer: Fixes for bidirectional searching and refactoring
At the cost of using more memory, now the forward and backward
searches use different caches for (position in image) -> (search node)

This is important so that we can allow the searches to overlap
before picking the best path.  At the moment this isn't very
sophisticated - we terminate on finding a newly explored point
is in the other search's closed list.  However, this fixes
obvious bugs with bidirectional searches in straight lines where
you'd sometimes get an unnecessary kink in the final trace.

This commit also removes a lot of repeated code and contains
assorted other bits of refactoring.