Simple_Neurite_Tracer: add a workaround for broken SWC files
Some of the SWC files distributed as part of the DIADEM Challenge [1]
inconveniently use indices into the image data (or "image coordinates")
instead of world coordinates in micrometers, as the specification
of the SWC format describes.

This is awkward to detect in general, and it's not clear then what
the radius field means in these files.

So, I have only added a workaround for these files, namely that if one
holds down the "Control" key when clicking on "Load Traces / SWC File"
and you select an SWC file, the loader will assume that the coordinates
are indexes into the image data rather than world coordinates, and
will assume that the radius is a multiple of the minimum voxel
separation (typically the same as the pixelWidth and pixelHeight)

In the future it would be good to warn the user and suggest this
workaround if the input data looks suspiciously as if the coordinates
are wrong.

[1] http://diademchallenge.org/